uploadthing error resolved

+++ src/app/api/uploadthing/core.ts
@@ -18,30 +18,37 @@ export const fileRouter = {
       return { user };
     })
     .onUploadComplete(async ({ metadata, file }) => {
-      const oldAvatarUrl = metadata.user.avatarUrl;
+      const dbUser = await prisma.user.findUnique({
+        where: { id: metadata.user.id },
+      });
+
+      if (!dbUser) {
+        throw new UploadThingError("User not found in database.");
+      }
+
+      const oldAvatarUrl = dbUser.avatarUrl;
 
       if (oldAvatarUrl) {
         const key = oldAvatarUrl.split(
           `/a/${process.env.NEXT_PUBLIC_UPLOADTHING_APP_ID}/`,
         )[1];
 
-        await new UTApi().deleteFiles(key);
+        if (key) {
+          await new UTApi().deleteFiles(key);
+        }
       }
 
-      const newAvatarUrl = file.url.replace(
-        "/f/",
-        `/a/${process.env.NEXT_PUBLIC_UPLOADTHING_APP_ID}/`,
-      );
+      const newAvatarUrl = file.url;
 
       await Promise.all([
         prisma.user.update({
-          where: { id: metadata.user.id },
+          where: { id: dbUser.id },
           data: {
             avatarUrl: newAvatarUrl,
           },
         }),
         streamServerClient.partialUpdateUser({
-          id: metadata.user.id,
+          id: dbUser.id,
           set: {
             image: newAvatarUrl,
           },
@@ -50,6 +57,7 @@ export const fileRouter = {
 
       return { avatarUrl: newAvatarUrl };
     }),
+
   attachment: f({
     image: { maxFileSize: "4MB", maxFileCount: 5 },
     video: { maxFileSize: "64MB", maxFileCount: 5 },
@@ -64,10 +72,7 @@ export const fileRouter = {
     .onUploadComplete(async ({ file }) => {
       const media = await prisma.media.create({
         data: {
-          url: file.url.replace(
-            "/f/",
-            `/a/${process.env.NEXT_PUBLIC_UPLOADTHING_APP_ID}/`,
-          ),
+          url: file.url,
           type: file.type.startsWith("image") ? "IMAGE" : "VIDEO",
         },
       });
